Optimize grid updates
instead of going through all cells to get their tile options, we keep a
cache of options, and only modify cells that might have changed.

Idea:
We start from the picked cell, and we check its neighbours. For each
neighbour (up, right, down, left), if the list of options has changed,
we add its own neighbours to the list of cells to check. If it didn't
change, we just skip its neighbours.

It is even better than going through the old grid because it propagates
changes until we come to a stable state. In the old way, we only check
cells once, according to the state at t-1, which can miss some option
reductions and lead to infeasibility.