Auto generate tile edge matching
When loading separate tiles, automatically detect which tile can be
adjacent to which other tiles by analyzing their edges.

- Added capability to load a single image and divide it in tiles in the
  tile-model code base.
- Extract the four edges of each tile and fuzzily match it with other
  edges.
- Make edges size calculation dependent on largest width of tiles.
- Allow applying a filter to the edge images to ignore some differences,
  for example a backgropund that should be ignored.
- Choose the edge size to be compatible with the small tiles generated
  from a single-image.
- Limit the comparison range to zero when the edge size is small.

Optimize the wave-function-collapse by using sets.

- Use set unions and intersections to update the grid.
- This is much faster when there are a lots of tiles.
- Change all options to be sets to use optimized set unions and intersections.
- Replaced up./down/left/right tile options with sets.
- Slightly optimize the set union by stopping when we have all tiles in
  the union.
- Limit the change propagation to a given radius to speed-up the
  algorithm.
- But take into account the case where we choose to collapse a cell that
  was not updated recently, which could lead to chosing a tile that
  should not be allowed. We always update the cell options before
  collapsing it to make sure it has updated options.

(We could also update its neighbors, etc... but we don't need to: if
they are already collapsed then they are valid. If they're not
collapsed, they will update themselves later on and if this leads to a
contradiction, then we will rewind.)

Add grid history and rewind to avoid restarting from scratch. This is in
the `histroy.js` file. Let the rewind to increase and decrease to allow
to go pass loops that always lead to a contradiction. By increasing the
rewind, we can go back far enough to remove the state that invariably
lead to a contradiction.

- This makes the tile WFC be optionally similar to the overlapping WFC.
- Added history of grids to be able to go back four steps instead of starting over.

Move the loading of tiles to its own file named tileLoader.js. This
allows different method of loading tiles and allow switching between
which tiles are used.

Split the draw function and updates into smaller functions.

Add debug function to help diagnose problems. Add debug logging to show
a cell tile and the options of its neighbor, to help debug when there
seems to be a contradiction. Debug code is in `debug.js`

Add UI to select the tiles to generate, restart the generation and show
tile neighbors options.